requirements.txt
pylibsalt>=0.1.0
UrwidMore>=0.1.0
pyreadpartitions>=1.0.0
